Refactor compiler dir structure
