+++ src/compiler/test-file/compiler-base.js
